refactor: extract our two custom remark into their own files
