+++ packages/gatsby-plugin-mdx/package.json
@@ -58,7 +58,6 @@
     "mdast-util-mdx": "^2.0.0",
     "mdast-util-to-markdown": "^1.3.0",
     "unified": "^10.1.2",
-    "unist-util-map": "^3.1.1",
     "unist-util-visit": "^4.1.0",
     "vfile": "^5.3.2"
   },
