+++ packages/gatsby-plugin-mdx/src/get-source-plugins-as-remark-plugins.ts
@@ -3,6 +3,8 @@ import type { GatsbyCache, Reporter, NodePluginArgs } from "gatsby"
 import type { Pluggable } from "unified"
 
 import { IMdxPluginOptions } from "./plugin-options"
+import { remarkMdxHtmlPlugin } from "./remark-mdx-html-plugin"
+import { pathPlugin } from "./remark-path-prefix-plugin"
 
 interface IGetSourcePluginsAsRemarkPlugins {
   gatsbyRemarkPlugins: IMdxPluginOptions["gatsbyRemarkPlugins"]
@@ -12,10 +14,6 @@ interface IGetSourcePluginsAsRemarkPlugins {
   reporter: Reporter
   cache: GatsbyCache
 }
-
-// ensure only one `/` in new url
-const withPathPrefix = (url: string, pathPrefix: string): string =>
-  (pathPrefix + url).replace(/\/\//, `/`)
 
 export async function getSourcePluginsAsRemarkPlugins({
   gatsbyRemarkPlugins,
@@ -27,35 +25,12 @@ export async function getSourcePluginsAsRemarkPlugins({
 }: IGetSourcePluginsAsRemarkPlugins): Promise<
   ProcessorOptions["remarkPlugins"]
 > {
-  const { visit } = await import(`unist-util-visit`)
-  const { map } = await import(`unist-util-map`)
-
-  // Ensure relative links include `pathPrefix`
-  const pathPlugin = () =>
-    async function transformer(markdownAST): Promise<any> {
-      if (!pathPrefix) {
-        return markdownAST
-      }
-      visit(markdownAST, [`link`, `definition`], node => {
-        if (
-          node.url &&
-          typeof node.url === `string` &&
-          node.url.startsWith(`/`) &&
-          !node.url.startsWith(`//`)
-        ) {
-          // TODO: where does withPathPrefix
-          node.url = withPathPrefix(node.url, pathPrefix)
-        }
-      })
-      return markdownAST
-    }
-
   const userPluginsFiltered = gatsbyRemarkPlugins.filter(
     plugin => typeof plugin.module === `function`
   )
 
   if (!userPluginsFiltered.length) {
-    return pathPrefix ? [pathPlugin] : []
+    return pathPrefix ? [[pathPlugin, { pathPrefix }]] : []
   }
 
   const userPlugins = userPluginsFiltered.map(plugin => {
@@ -81,58 +56,6 @@ export async function getSourcePluginsAsRemarkPlugins({
             },
             plugin.options || {}
           )
-
-          const divAST = map(markdownAST, node => {
-            if (node.type !== `html`) {
-              return node
-            }
-            return {
-              type: `mdxJsxFlowElement`,
-              name: `div`,
-              attributes: [
-                {
-                  type: `mdxJsxAttribute`,
-                  name: `dangerouslySetInnerHTML`,
-                  value: {
-                    type: `mdxJsxAttributeValueExpression`,
-                    data: {
-                      estree: {
-                        type: `Program`,
-                        body: [
-                          {
-                            type: `ExpressionStatement`,
-                            expression: {
-                              type: `ObjectExpression`,
-                              properties: [
-                                {
-                                  type: `Property`,
-                                  method: false,
-                                  shorthand: false,
-                                  computed: false,
-                                  key: {
-                                    type: `Identifier`,
-                                    name: `__html`,
-                                  },
-                                  value: {
-                                    type: `Literal`,
-                                    value: node.value,
-                                  },
-                                  kind: `init`,
-                                },
-                              ],
-                            },
-                          },
-                        ],
-                        sourceType: `module`,
-                      },
-                    },
-                  },
-                },
-              ],
-            }
-          })
-
-          return divAST
         }
       }
 
@@ -140,8 +63,8 @@ export async function getSourcePluginsAsRemarkPlugins({
   })
 
   if (pathPrefix) {
-    return [pathPlugin, ...userPlugins]
+    return [[pathPlugin, { pathPrefix }], ...userPlugins, remarkMdxHtmlPlugin]
   } else {
-    return userPlugins
+    return [...userPlugins, remarkMdxHtmlPlugin]
   }
 }
